extract core interfaces
https://coveord.atlassian.net/browse/KIT-632
